+++ packages/bezier-codemod/src/App.tsx
@@ -18,16 +18,16 @@ import {
 } from 'ink'
 
 import project from './project.js'
-import enumMemberToStringLiteral from './transforms/enum-member-to-string-literal/enum-member-to-string-literal.js'
-import foundationToCssVariableBorder from './transforms/foundation-to-css-variable/foundation-to-css-variable-border.js'
-import foundationToCssVariableElevation from './transforms/foundation-to-css-variable/foundation-to-css-variable-elevation.js'
-import foundationToCssVariableRounding from './transforms/foundation-to-css-variable/foundation-to-css-variable-rounding.js'
-import foundationToCssVariableTheme from './transforms/foundation-to-css-variable/foundation-to-css-variable-theme.js'
-import foundationToCssVariableTransition from './transforms/foundation-to-css-variable/foundation-to-css-variable-transition.js'
-import foundationToCssVariable from './transforms/foundation-to-css-variable/foundation-to-css-variable.js'
-import iconNameToBezierIcon from './transforms/icon-name-to-bezier-icons/icon-name-to-bezier-icon.js'
-import iconsToBezierIcons from './transforms/icons-to-bezier-icons/icons-to-bezier-icons.js'
-import styledToStyledComponents from './transforms/import-styled-from-styled-components/import-styled-from-styled-component.js'
+import enumMemberToStringLiteral from './transforms/enum-member-to-string-literal/transform.js'
+import foundationToCssVariableBorder from './transforms/foundation-to-css-variable/border.js'
+import foundationToCssVariableElevation from './transforms/foundation-to-css-variable/elevation.js'
+import foundationToCssVariableRounding from './transforms/foundation-to-css-variable/rounding.js'
+import foundationToCssVariableTheme from './transforms/foundation-to-css-variable/theme.js'
+import foundationToCssVariable from './transforms/foundation-to-css-variable/transform.js'
+import foundationToCssVariableTransition from './transforms/foundation-to-css-variable/transition.js'
+import iconNameToBezierIcon from './transforms/icon-name-to-bezier-icons/transform.js'
+import iconsToBezierIcons from './transforms/icons-to-bezier-icons/transform.js'
+import styledToStyledComponents from './transforms/import-styled-from-styled-components/transform.js'
 import interpolationToCssVariable from './transforms/interpolation-to-css-variable.js'
 import removeAlphaFromAlphaStack from './transforms/remove-alpha-from-alpha-stack.js'
