+++ packages/bezier-codemod/src/transforms/tests/import-styled-from-styled-components/import-styled-from-styled-components.test.ts
@@ -1,5 +1,5 @@
 import { testTransformFunction } from '../../../utils/test.js'
-import importTransform from '../../import-styled-from-styled-components/import-styled-from-styled-component.js'
+import importTransform from '../../import-styled-from-styled-components/transform.js'
 
 describe('import transform', () => {
   it('should transform styled import from bezier-react to styled-components (1)', () => {
